+++ core/src/main/java/org/seedstack/seed/core/internal/crypto/DecryptFunction.java
@@ -5,12 +5,14 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
+
 package org.seedstack.seed.core.internal.crypto;
 
 import com.google.common.base.Strings;
 import com.google.common.io.BaseEncoding;
 import java.nio.charset.StandardCharsets;
 import java.security.KeyStore;
+import java.util.concurrent.atomic.AtomicBoolean;
 import org.seedstack.coffig.Coffig;
 import org.seedstack.coffig.spi.ConfigFunction;
 import org.seedstack.coffig.spi.ConfigFunctionHolder;
@@ -20,21 +22,28 @@ import org.seedstack.seed.crypto.CryptoConfig;
 import org.seedstack.seed.crypto.EncryptionService;
 
 public class DecryptFunction implements ConfigFunctionHolder {
+    private final AtomicBoolean initInProgress = new AtomicBoolean(false);
     private EncryptionServiceFactory encryptionServiceFactory;
     private CryptoConfig.KeyStoreConfig masterKeyStoreConfig;
     private Exception storedException;
 
     @Override
     public void initialize(Coffig coffig) {
-        CryptoConfig cryptoConfig = coffig.get(CryptoConfig.class);
-        masterKeyStoreConfig = cryptoConfig.masterKeyStore();
-        if (masterKeyStoreConfig != null) {
-            try {
-                KeyStore keyStore = new KeyStoreLoader().load(CryptoConfig.MASTER_KEY_STORE_NAME, masterKeyStoreConfig);
-                encryptionServiceFactory = new EncryptionServiceFactory(cryptoConfig, keyStore);
-            } catch (Exception e) {
-                storedException = e;
+        try {
+            initInProgress.set(true);
+            CryptoConfig cryptoConfig = coffig.get(CryptoConfig.class);
+            masterKeyStoreConfig = cryptoConfig.masterKeyStore();
+            if (masterKeyStoreConfig != null) {
+                try {
+                    KeyStore keyStore = new KeyStoreLoader().load(CryptoConfig.MASTER_KEY_STORE_NAME,
+                            masterKeyStoreConfig);
+                    encryptionServiceFactory = new EncryptionServiceFactory(cryptoConfig, keyStore);
+                } catch (Exception e) {
+                    storedException = e;
+                }
             }
+        } finally {
+            initInProgress.set(false);
         }
     }
 
@@ -45,19 +54,24 @@ public class DecryptFunction implements ConfigFunctionHolder {
 
     @ConfigFunction
     String decrypt(String alias, String value) {
-        if (encryptionServiceFactory == null) {
-            if (storedException != null) {
-                throw SeedException.wrap(storedException, CryptoErrorCode.MISSING_MASTER_KEYSTORE);
-            } else {
-                throw SeedException.createNew(CryptoErrorCode.MISSING_MASTER_KEYSTORE);
+        if (initInProgress.get()) {
+            // Cannot decrypt anything during initialization phase
+            return value;
+        } else {
+            if (encryptionServiceFactory == null) {
+                if (storedException != null) {
+                    throw SeedException.wrap(storedException, CryptoErrorCode.MISSING_MASTER_KEYSTORE);
+                } else {
+                    throw SeedException.createNew(CryptoErrorCode.MISSING_MASTER_KEYSTORE);
+                }
             }
+            CryptoConfig.KeyStoreConfig.AliasConfig aliasConfig = masterKeyStoreConfig.getAliases().get(alias);
+            if (aliasConfig == null || Strings.isNullOrEmpty(aliasConfig.getPassword())) {
+                throw SeedException.createNew(CryptoErrorCode.MISSING_MASTER_KEY_PASSWORD);
+            }
+            EncryptionService encryptionService = encryptionServiceFactory.create(alias,
+                    aliasConfig.getPassword().toCharArray());
+            return new String(encryptionService.decrypt(BaseEncoding.base16().decode(value)), StandardCharsets.UTF_8);
         }
-        CryptoConfig.KeyStoreConfig.AliasConfig aliasConfig = masterKeyStoreConfig.getAliases().get(alias);
-        if (aliasConfig == null || Strings.isNullOrEmpty(aliasConfig.getPassword())) {
-            throw SeedException.createNew(CryptoErrorCode.MISSING_MASTER_KEY_PASSWORD);
-        }
-        EncryptionService encryptionService = encryptionServiceFactory.create(alias,
-                aliasConfig.getPassword().toCharArray());
-        return new String(encryptionService.decrypt(BaseEncoding.base16().decode(value)), StandardCharsets.UTF_8);
     }
 }