+++ rest/jersey2/src/test/java/org/seedstack/seed/rest/jersey2/fixtures/Something.java
@@ -5,12 +5,13 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
+
 package org.seedstack.seed.rest.jersey2.fixtures;
 
 import javax.validation.Valid;
 import javax.validation.constraints.Max;
 import javax.validation.constraints.Min;
-import javax.validation.constraints.NotBlank;
+import org.seedstack.seed.validation.NotBlank;
 
 public class Something {
     @NotBlank(message = "someI18nKey")