+++ rest/jersey2/src/test/java/org/seedstack/seed/rest/jersey2/fixtures/ValidatingResource.java
@@ -5,16 +5,17 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
+
 package org.seedstack.seed.rest.jersey2.fixtures;
 
 import javax.validation.Valid;
-import javax.validation.constraints.NotBlank;
 import javax.ws.rs.GET;
 import javax.ws.rs.POST;
 import javax.ws.rs.Path;
 import javax.ws.rs.QueryParam;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.core.UriBuilder;
+import org.seedstack.seed.validation.NotBlank;
 
 @Path("/validating")
 public class ValidatingResource {