+++ specs/src/main/java/org/seedstack/seed/validation/NotBlankValidator.java
@@ -1,0 +1,40 @@
+/*
+ * Copyright © 2013-2019, The SeedStack authors <http://seedstack.org>
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package org.seedstack.seed.validation;
+
+import javax.validation.ConstraintValidator;
+import javax.validation.ConstraintValidatorContext;
+
+public class NotBlankValidator implements ConstraintValidator<NotBlank, String> {
+    @Override
+    public void initialize(NotBlank constraintAnnotation) {
+        // no-op
+    }
+
+    @Override
+    public boolean isValid(String value, ConstraintValidatorContext context) {
+        if (value == null) {
+            return false;
+        } else {
+            int length = value.length();
+            return indexOfNonWhitespace(value, length) < length;
+        }
+    }
+
+    private int indexOfNonWhitespace(String value, int length) {
+        int left = 0;
+        while (left < length) {
+            char ch = value.charAt(left);
+            if (ch != ' ' && ch != '\t' && !Character.isWhitespace(ch)) {
+                break;
+            }
+            left++;
+        }
+        return left;
+    }
+}