+++ web/undertow/src/main/java/org/seedstack/seed/undertow/internal/DeploymentManagerFactory.java
@@ -74,7 +74,7 @@ class DeploymentManagerFactory {
         }
 
         for (ServletContainerInitializer servletContainerInitializer : loadServletContainerInitializers()) {
-            deploymentInfo.addServletContainerInitalizer(
+            deploymentInfo.addServletContainerInitializer(
                     createServletContainerInitializerInfo(servletContainerInitializer));
         }